Rename hook and use combine queries
