+++ packages/web/app/api/queries/index.js
@@ -12,4 +12,4 @@ export * from './useTranslations';
 export * from './useTranslationLanguages';
 export * from './useTelegramBotInfoQuery';
 export * from './useChartSurveysQuery';
-export * from './useChartQuery';
+export * from './useEncounterChartsQuery';
